requirements.txt
py-cord
pandas
openpyxl
requests
python-dotenv
apscheduler
PyNaCl